Moved  MeasurementDataForEndpoint and MeasurementData to electrical-energy-measurement-server.h
