+++ src/app/clusters/electrical-energy-measurement-server/electrical-energy-measurement-server.h
@@ -26,6 +26,16 @@ namespace app {
 namespace Clusters {
 namespace ElectricalEnergyMeasurement {
 
+struct MeasurementData
+{
+    Structs::MeasurementAccuracyStruct::Type measurementAccuracy;
+    Optional<Structs::EnergyMeasurementStruct::Type> cumulativeImported;
+    Optional<Structs::EnergyMeasurementStruct::Type> cumulativeExported;
+    Optional<Structs::EnergyMeasurementStruct::Type> periodicImported;
+    Optional<Structs::EnergyMeasurementStruct::Type> periodicExported;
+    Optional<Structs::CumulativeEnergyResetStruct::Type> cumulativeReset;
+};
+
 bool NotifyCumulativeEnergyMeasured(EndpointId endpointId, const Optional<Structs::EnergyMeasurementStruct::Type> & energyImported,
                                     const Optional<Structs::EnergyMeasurementStruct::Type> & energyExported);
 
@@ -35,6 +45,8 @@ bool NotifyPeriodicEnergyMeasured(EndpointId endpointId, const Optional<Structs:
 CHIP_ERROR SetMeasurementAccuracy(EndpointId endpointId, const Structs::MeasurementAccuracyStruct::Type & accuracy);
 
 CHIP_ERROR SetCumulativeReset(EndpointId endpointId, const Optional<Structs::CumulativeEnergyResetStruct::Type> & cumulativeReset);
+
+MeasurementData * MeasurementDataForEndpoint(EndpointId endpointId);
 
 } // namespace ElectricalEnergyMeasurement
 } // namespace Clusters
